+++ src/main/java/DataParseCoordinates.java
@@ -6,10 +6,16 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
-public class DataParsingCountry {
-    public static String getter() throws IOException {
+
+public class DataParseCoordinates {
+    public static String getter(String city) throws IOException {
         String key = null;
-        String urlString = "http://htmlweb.ru/geo/api.php?locations&json&api_key=API_KEY_из_профиля";
+        try {
+            key = new String(Commands.class.getResourceAsStream("/keyCity.txt").readAllBytes(), StandardCharsets.UTF_8);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        String urlString = "https://api.opencagedata.com/geocode/v1/json?q=" + city + "&key=" + key;
         URL url = new URL(urlString);
         HttpURLConnection connection = (HttpURLConnection) url.openConnection();
         connection.setRequestMethod("GET");
@@ -24,7 +30,6 @@ public class DataParsingCountry {
             }
             reader.close();
             String responseData = response.toString();
-            //System.out.println("Полученные данные: " + responseData);
             return toJson(responseData);
 
         } else {
@@ -37,6 +42,11 @@ public class DataParsingCountry {
     public static String toJson(String request) throws IOException {
         JsonParser parser = new JsonParser();
         JsonObject jsonObject = parser.parse(request).getAsJsonObject();
-        return "0";
+        JsonArray resultsArray = jsonObject.getAsJsonArray("results");
+        JsonObject firstElement  = resultsArray.get(0).getAsJsonObject();
+        JsonObject geometryObject = firstElement.getAsJsonObject("geometry");
+        String lat = String.valueOf(geometryObject.get("lat")); // Широта
+        String lng = String.valueOf(geometryObject.get("lng")); // Долгота
+        return DataParseWeather.getter(lat, lng);
     }
 }
